rename faq api elements
